backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
openai
requests